requirements.txt
backcall==0.1.0
decorator==4.4.0
Django==2.2
django-extensions==2.1.9
django-gm2m==0.6.1
djangorestframework==3.9.4
djangorestframework-gis==0.14
entrypoints==0.3
flake8==3.7.7
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
mccabe==0.6.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
psycopg2-binary==2.8.2
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.2
pygraphviz==1.5
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
wcwidth==0.1.7
